// src/app/models/hospital.ts
export class Hospital {
  id: number;
  name: string;
  description: string;
  metaKeyword: string;
  metaDescription: string;
  address: string;
  email: string;
  phone: string;
  logo: string;
  favicon: string;
  backgroundSlider: string;
  bannerContentHeader: string;
  bannerContentParagraph: string;
  copyright: string;
  twitterApi: string;
  googleMap: string;
  facebookUrl: string;
  twitterUrl: string;
  instagramUrl: string;
  googlePlusUrl: string;
  dribbbleUrl: string;
  skypeUrl: string;
  vimeoUrl: string;
  status: Number = 0;
}